% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actfts-package.R
\docType{package}
\name{actfts-package}
\alias{actfts}
\alias{actfts-package}
\title{actfts: Autocorrelation Tools Featured for Time Series}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package provides tools for performing autocorrelation analysis of time series data. It includes functions to compute and visualize the autocorrelation function (ACF) and the partial autocorrelation function (PACF). Additionally, it performs the Dickey-Fuller, KPSS and Phillips–Perron unit root tests to assess the stationarity of time series.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SergioFinances/actfts}
  \item \url{https://sergiofinances.github.io/actfts/}
  \item Report bugs at \url{https://github.com/SergioFinances/actfts/issues}
}

}
\author{
\strong{Maintainer}: Sergio Sierra \email{sergiochess95@gmail.com} (\href{https://orcid.org/0000-0003-4055-8810}{ORCID})

Authors:
\itemize{
  \item David Rodríguez \email{davestss@hotmail.com} (\href{https://orcid.org/0000-0001-5430-0787}{ORCID}) [copyright holder]
}

}
\keyword{internal}
