% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPIEEUU.R
\docType{data}
\name{DPIEEUU}
\alias{DPIEEUU}
\title{Gross Domestic Product of the United States.}
\format{
A dataset in xts format with one variable and several files that are automatically updated depending on the update of the data's FRED:
\describe{
  \item{DPI}{Disposable Personal Income in billions of dollars}
}
}
\source{
\url{https://fred.stlouisfed.org/series/DPI}
}
\usage{
DPIEEUU
}
\description{
This dataset contains the disposable personal income of the United States from 01/01/1947 to the present with quartely frequency.
}
\references{
U.S. Bureau of Economic Analysis, Disposable Personal Income (DPI), retrieved from FRED, Federal Reserve Bank of St. Louis; https://fred.stlouisfed.org/series/DPI
}
\keyword{datasets}
