% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDPEEUU.R
\docType{data}
\name{GDPEEUU}
\alias{GDPEEUU}
\title{Gross Domestic Product of the United States.}
\format{
A dataset in xts format with one variable and several files that are automatically updated depending on the update of the data's FRED:
\describe{
\item{GDP}{Gross Domestic product in billions of dollars}
}
}
\source{
\url{https://fred.stlouisfed.org/series/GDP}
}
\usage{
GDPEEUU
}
\description{
This dataset contains the gross domestic product of the United States from 01/01/1947 to the present with quartely frequency.
}
\references{
U.S. Bureau of Economic Analysis, Gross Domestic Product (GDP), retrieved from FRED, Federal Reserve Bank of St. Louis; https://fred.stlouisfed.org/series/GDP
}
\keyword{datasets}
