% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCECEEUU.R
\docType{data}
\name{PCECEEUU}
\alias{PCECEEUU}
\title{Personal Consumption Expenditures of the United States.}
\format{
A dataset in xts format with one variable and several files that are automatically updated depending on the update of the data's FRED:
\describe{
\item{PCEC}{Personal Consumption Expenditures in billions of dollars}
}
}
\source{
\url{https://fred.stlouisfed.org/series/PCEC}
}
\usage{
PCECEEUU
}
\description{
This dataset contains the personal consumption expenditures of the United States from 01/01/1947 to the present with quartely frequency.
}
\references{
U.S. Bureau of Economic Analysis, Personal Consumption Expenditures (PCEC), retrieved from FRED, Federal Reserve Bank of St. Louis; https://fred.stlouisfed.org/series/PCEC
}
\keyword{datasets}
